refactor(312 C1): extract ChatMessage VO + helpers from session.py → runtime/chat_message.py
FP-0044 C-series stage 1 (god-file decomposition). Move the ChatMessage value
object + its read-time legacy migration (_migrate_legacy_chat_message) + the
_now_iso timestamp helper out of the session.py god-file into their own module.
Byte-identical class/helper bodies (verbatim). Pure VO — no Session dependency
(Session imports chat_message; one-directional, no cycle). Class name unchanged
(the "chat" legacy naming is a separate optional cleanup, per lead).

- New runtime/chat_message.py: ChatMessage (@dataclass(init=False)) +
  _migrate_legacy_chat_message + _now_iso (+ dataclasses/datetime/typing imports).
- session.py: removed the 165-line region; imports the 3 symbols from
  chat_message for internal use. God-file shrinks 5214 → 5049 LOC.
- runtime/__init__: lazy ChatMessage re-export → .chat_message.
- Repointed 22 consumers (+ lazy + TYPE_CHECKING): the 3 moved symbols import
  from chat_message; Session/RouterCapExceeded/etc. stay on session (mixed
  imports split). Repo-wide; straggler 0.

Verification: collect-only 7581 (imports resolve); no cycle; ChatMessage smoke
(text + agent-reject). Consumer sweep 1249 passed; the 1 fail
(test_web_fetch_preview_path_ref legacy-no-media-store, HTML extraction) is
pre-existing local-env (reproduces on clean main with this stashed — falsified).
ruff clean.

Refs #312
