+++ tests/test_session_notify_state_change.py
@@ -28,7 +28,8 @@ from pathlib import Path
 import pytest
 
 from reyn.core.events.state_log import StateLog
-from reyn.runtime.session import ChatMessage, Session
+from reyn.runtime.chat_message import ChatMessage
+from reyn.runtime.session import Session
 
 
 def _make_session(tmp_path: Path, *, agent_name: str = "alpha") -> Session:
