+++ tests/test_session_router_history_slicing.py
@@ -24,7 +24,8 @@ import reyn.llm.model_budget as _mb
 from reyn.config import CompactionConfig
 from reyn.core.events.state_log import StateLog
 from reyn.runtime.budget.budget import BudgetTracker, CostConfig
-from reyn.runtime.session import ChatMessage, Session
+from reyn.runtime.chat_message import ChatMessage
+from reyn.runtime.session import Session
 
 
 def _now() -> str:
